crashreport: SWUPDATE changed to INFO event from REBOOT event
because SWUPDATE was separated from startup reason and changed
to an INFO event rather than a REBOOT event, crashreport needed
to be updated for data ready management.

Tracked-On: GMINL-11295
Reviewed-on: 381119
Tested-by: Kumar-mayernik, Nisha <nisha.kumar-mayernik@intel.com>

+++ crashreport/src/com/intel/crashreport/specific/CheckEventsService.java
@@ -238,12 +238,12 @@ public class CheckEventsService extends Service {
 									else if (ret == -3)
 										Log.w(from+": Event " +event.toString() + " with wrong date, addition in DB canceled");
 									else {
+										if (event.getType().contentEquals("SWUPDATE") &&
+											event.getEventName().contentEquals("INFO")){
+											db.deleteEventsBeforeUpdate(event.getEventId());
+										}
 										if (event.getEventName().contentEquals("REBOOT")) {
-											if (event.getType().contentEquals("SWUPDATE")){
-												db.deleteEventsBeforeUpdate(event.getEventId());
-											}else{
-												db.updateEventsNotReadyBeforeREBOOT(event.getEventId());
-											}
+											db.updateEventsNotReadyBeforeREBOOT(event.getEventId());
 										}
 										if (event.getEventName().equals("BZ")) {
 											try {
